Fixes overtone circle scaling coordinates issue. Fixes #43.
Chrome decided to change its default behavior for SVGs to the same as
Firefox. Fortunately, it seems that support for `transform-box: fill-box`
has arrived in both browser, fixing the issue on Chrome and resolving
the browser specific problem with Firefox. Older versions of Firefox might
now see something weird, but that's probably ok.